Update FerCorrales.cs
You just merged my pull request and by visiting Planet Powershell authors page, I noticed there was a small typo. I am trying to get it fixed, it is something minor but would be good to change it. Sorry about that and thanks so much for adding me, and for helping get this corrected.