Update FerCorrales.cs
Thanks for adding me. You just merged my pull request and by visiting Planet Powershell authors page, I noticed there was a small typo which I am trying to get fixed, it is something minor but would be good to change it. I would also like to take the opportunity to replace "PS" with "PowerShell". Sorry about that and thanks again for accepting me as a featured author, and for helping get this corrected.